requirements.txt
# Django and core dependencies
Django>=5.2,<6.0
asgiref>=3.8.1
sqlparse>=0.3.1
python-dotenv>=1.0.0
django-cors-headers>=4.0.0

# PDF processing
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
pytesseract>=0.3.10

# AI and ML
openai>=1.0.0
yake>=0.4.8
langdetect>=1.0.9
rapidfuzz>=3.0.0

# Production server
gunicorn>=21.0.0
whitenoise>=6.0.0

# Database (for production)
psycopg2-binary>=2.9.0

# Security
django-cors-headers>=4.0.0

# Monitoring and logging
sentry-sdk[django]>=1.0.0

# OpenTelemetry observability
opentelemetry-api>=1.36.0
opentelemetry-sdk>=1.36.0
opentelemetry-instrumentation-django>=0.57b0
opentelemetry-instrumentation-psycopg2>=0.57b0
opentelemetry-instrumentation-requests>=0.57b0
opentelemetry-instrumentation-sqlalchemy>=0.57b0
opentelemetry-instrumentation-celery>=0.57b0
opentelemetry-instrumentation-redis>=0.57b0
opentelemetry-exporter-jaeger>=1.21.0
opentelemetry-exporter-prometheus>=0.57b0
opentelemetry-exporter-otlp>=1.15.0

# System dependencies (will be installed in Docker)
# tesseract-ocr
# poppler-utils